refactor: Move rendering to freya-core
