+++ crates/core/src/dom/mod.rs
@@ -1,7 +1,6 @@
 pub mod dom_adapter;
 pub mod doms;
 mod mutations_writer;
-mod paragraph_utils;
 
 pub use dom_adapter::*;
 pub use doms::*;
