+++ crates/core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod node;
 pub mod platform_state;
 pub mod plugins;
 pub mod render;
+pub mod skia;
 pub mod types;
 
 pub mod prelude {
@@ -18,6 +19,7 @@ pub mod prelude {
         platform_state::*,
         plugins::*,
         render::*,
+        skia::*,
         types::*,
     };
 }
