+++ crates/core/src/skia/paragraph.rs
@@ -5,15 +5,19 @@ use freya_common::{
     CursorLayoutResponse,
     TextGroupMeasurement,
 };
-use freya_native_core::real_dom::NodeImmutable;
+use freya_native_core::prelude::NodeImmutable;
 use freya_node_state::CursorSettings;
-use torin::prelude::*;
-
-use crate::{
-    dom::DioxusNode,
-    layout::align_main_align_paragraph,
+use torin::prelude::{
+    CursorPoint,
+    LayoutNode,
 };
 
+use crate::prelude::{
+    align_main_align_paragraph,
+    DioxusNode,
+};
+
+/// Merasure the cursor positio and text selection and notify the subscribed component of the element.
 pub fn measure_paragraph(
     node: &DioxusNode,
     layout_node: &LayoutNode,
