+++ crates/core/src/skia/skia_renderer/image.rs
@@ -1,4 +1,3 @@
-use freya_core::dom::DioxusNode;
 use freya_engine::prelude::*;
 use freya_native_core::real_dom::NodeImmutable;
 use freya_node_state::{
@@ -6,6 +5,8 @@ use freya_node_state::{
     Style,
 };
 use torin::geometry::Area;
+
+use crate::dom::DioxusNode;
 
 /// Render an `image` element
 pub fn render_image(area: &Area, node_ref: &DioxusNode, canvas: &Canvas) {
