+++ crates/core/src/skia/skia_renderer/label.rs
@@ -1,13 +1,14 @@
 use std::sync::Arc;
 
 use freya_common::CachedParagraph;
-use freya_core::prelude::{
-    align_main_align_paragraph,
-    DioxusNode,
-};
 use freya_engine::prelude::*;
 use freya_native_core::prelude::SendAnyMap;
 use torin::geometry::Area;
+
+use crate::prelude::{
+    align_main_align_paragraph,
+    DioxusNode,
+};
 
 /// Render a `label` element
 pub fn render_label(
