+++ crates/core/src/skia/skia_renderer/mod.rs
@@ -1,0 +1,9 @@
+mod image;
+mod label;
+mod paragraph;
+mod rect;
+mod renderer;
+mod svg;
+mod wireframe;
+
+pub use renderer::*;
