--- crates/renderer/src/elements/mod.rs
@@ -1,11 +1,0 @@
-mod image;
-mod label;
-mod paragraph;
-mod rect;
-mod svg;
-
-pub use image::*;
-pub use label::*;
-pub use paragraph::*;
-pub use rect::*;
-pub use svg::*;
