+++ crates/renderer/src/lib.rs
@@ -14,10 +14,7 @@ mod accessibility;
 mod app;
 mod config;
 pub mod devtools;
-mod elements;
-mod render;
 mod renderer;
 mod winit_waker;
-mod wireframe;
 
 pub type HoveredNode = Option<Arc<Mutex<Option<NodeId>>>>;
