+++ crates/state/tests/parse_highlight.rs
@@ -1,4 +1,3 @@
-use freya_engine::prelude::*;
 use freya_node_state::{
     HighlightMode,
     Parse,
